26th

+++ src/helpers/axios.js
@@ -17,14 +17,14 @@ axiosIntance.interceptors.request.use((req) => {
   }
   return req;
 });
-axiosIntance.interceptors.request.use(
+axiosIntance.interceptors.response.use(
   (res) => {
     return res;
   },
   (error) => {
     console.log(error);
-    const { status } = error.response;
-    if (status === 500) {
+    const status = error.response ? error.response.status : 500;
+    if (status && status === 500) {
       localStorage.clear();
       store.dispatch({ type: authConstants.LOGOUT_SUCCESS });
     }
